tests: combine library & library item tests
* test_list_create_libraries.py
* test_update_destroy_library.py
* test_create_library_item.py
* test_update_destroy_library_item.py

becomes test_libraries.py
